+++ pkg/nificlient/client.go
@@ -102,6 +102,7 @@ type NifiClient interface {
 	GetLabel(id string) (*nigoapi.LabelEntity, error)
 	CreateLabel(entity nigoapi.LabelEntity, pgParentId string) (*nigoapi.LabelEntity, error)
 	UpdateLabel(entity nigoapi.LabelEntity) (*nigoapi.LabelEntity, error)
+	RemoveLabel(entity nigoapi.LabelEntity) error
 
 	// User groups func
 	GetUserGroups() ([]nigoapi.UserGroupEntity, error)